requirements.txt
PyMySQL==0.9.3
SQLAlchemy==1.3.12
web3==5.4.0
python-telegram-bot
requests